move root level getLogger calls to actually read from the env properly, and small fixes for env and env validation
